Update followMe.py
Fixed as many code-breaking bugs as possible, completely scratched checkLeftRight() and rewrote, made functions that enable the user to store and call the servo position, pretty much done! Only thing left to do is teach the robot to dance and fix the little things. Coming in on the home stretch now.